requirements.txt
Django
google-auth-oauthlib
google-auth
oauthlib
requests-oauthlib
requests
django-allauth
pyjwt
cryptography
